photo-renamer/requirements.txt
pillow==10.2.0  # Image processing and EXIF data reading
pyyaml==6.0.1  # YAML configuration file parsing
python-dotenv==1.0.0  # Environment variable management
